requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
iso8601==1.0.2
multidict==6.0.2
pip==21.1.2
python-dotenv==0.20.0
requests==2.27.1
setuptools==57.0.0
twitchio==2.2.0
typing-extensions==4.1.1
urllib3==1.26.9
wheel==0.36.2
yarl==1.7.2
